fix all error

+++ src/app/modules/user/user.service.ts
@@ -13,6 +13,10 @@ const getUsers = async (): Promise<IUser[]> => {
   const result = await User.find({}).select(
     'userId username fullName.firstName fullName.lastName age email address.street address.city address.country',
   );
+  console.log(result);
+  if (!result.length) {
+    throw userNotFound('No users found', 404, 'No users found!');
+  }
   return result;
 };
 //get single user
@@ -23,7 +27,7 @@ const getUser = async (userId: number): Promise<IUser | null> => {
     const result = await User.findOne({ userId });
     return result;
   } else {
-    throw userNotFound();
+    throw userNotFound('User not found', 404, 'User not found!');
   }
 };
 //Update user
@@ -31,13 +35,23 @@ const updateUser = async (
   userId: number,
   data: Partial<IUser>,
 ): Promise<IUser | null> => {
-  const result = await User.findOneAndUpdate({ userId }, data, { new: true });
-  return result;
+  console.log(await User.isUserExists(userId));
+
+  if (await User.isUserExists(userId)) {
+    const result = await User.findOneAndUpdate({ userId }, data, { new: true });
+    return result;
+  } else {
+    throw userNotFound('User not found', 404, 'User not found!');
+  }
 };
 // Delete user
 const deleteUser = async (userId: number): Promise<IUser | null> => {
-  const result = await User.findOneAndDelete({ userId });
-  return result;
+  if (await User.isUserExists(userId)) {
+    const result = await User.findOneAndDelete({ userId });
+    return result;
+  } else {
+    throw userNotFound('User not found', 404, 'User not found!');
+  }
 };
 export const UserServices = {
   createUser,
